Use proper @import rules.
Fixes: Sass::SyntaxError: File to import not found or unreadable: icon-font.

I implemented a `replace_file_imports` similiar to bootstrap-sass's. These imports are now fully compliant with the proper path/namespace to use. In my example this fixes all imports like this.

```less
@import "icon-font";
```

To this:

```sass
@import "flat-ui-pro/icon-font";
```

This follows the convention of an asset gem using the directory for the namespace and is exactly what bootstrap-sass does too.